+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # NHS.UK frontend Changelog
 
+## Unreleased
+
+:new: **New features**
+- Add `role="button"` to links _visually styled_ as buttons (`<a class="nhsuk-button" role="button">`) and bind a spacebar keydown listener to these elements so that they _behave_ like buttons.
+- Add simple "debounce" behaviour to buttons to catch double clicks and prevent double submissions.
+
 ## 6.1.2 - 8 August 2022
 
 :wrench: **Fixes**